Update AnnotationManagerBean to properly handle documents with multiple internal ids (#3260)
The major change here is that `lookupDocumentIdentifier` in
AnnotationManagerBean now returns a `List<Metadata>` instead of a
`Metadata` objects, so all bean methods had to be updated to work with
more than one `Metadata` object.

Also implemented a separate annotation-sepecific `Metadata` different
from the one in the event package because the two contexts has different
needs and it made sense to decouple the two.